Extract EGM & geomorph import into MorphAnimation
